requirements.txt
Django==2.0.4
django-shortuuidfield==0.1.3
pytz==2018.4
reverse==0.1.0
shortuuid==0.5.0
six==1.11.0
